refactor(jarvis): split server.py's 44 models + package discovery into owner modules
clio-kit campaign #362 wave 1, Slice 1: server.py was 3109 lines carrying 44
Pydantic/TypedDict/dataclass model classes plus the tool-registration surface
plus the package-discovery/search implementation. Splits it into:

- jarvis_mcp/models/ (compat, packages, describe, execution, progress,
  artifact_documents, datasets, service_runtime): every wire-facing document,
  grouped by concern.
- jarvis_mcp/package_discovery.py: package inventory discovery, bounded
  search, and agent-visible settings extraction (previously inline in
  server.py). Its one JARVIS-manager-singleton call is a lazily-resolved
  import from server.py to avoid a cycle while keeping
  unittest.mock.patch("jarvis_mcp.server.get_manager", ...) working exactly
  as before.
- server.py: 3109 -> 1323 lines, now just FastMCP init, the `@mcp.tool`
  registration surface, the manager singleton, and CLI entry points.

Two test patch targets (_package_agent_metadata,
_package_configuration_search_text) moved with their implementation to
jarvis_mcp.package_discovery; every other jarvis_mcp.server.* patch/import
site tests relied on is unchanged (verified against the full grep of every
patch("jarvis_mcp.server....") and "from jarvis_mcp.server import ..." call
in the suite before touching any code).

The wire surface is unchanged: tests/test_contract_pins.py (committed before
this split) passes byte-for-byte against the split code, and
scripts/validate_fastmcp.py / the server.json tool-name diff both pass.
Full suite: 424 passed (419 baseline + 5 new contract pins), ruff clean,
mypy clean.
